Move memcache update logic out of REST handler
